BreedSenseAI-main/requirements.txt
Flask==3.1.0
pymongo==4.11.3
torch==2.6.0
torchvision==0.21.0
Pillow==11.1.0
python-dotenv==1.0.1
gunicorn==23.0.0
flask-cors==5.0.1
matplotlib==3.10.0
numpy==2.2.3
flask-login==0.6.3
bcrypt==4.2.1
